+++ src/ndr_asttokeninformation.h
@@ -44,7 +44,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 //typedef struct NDR_TreeTokenInfoWrapper NDR_TreeTokenInfoWrapper;
 
 typedef struct NDR_TreeTokenInfo{
-    size_t nodeNumber;
+    long nodeNumber;
     NDR_TokenInformation* tokenInfo;
 } NDR_TreeTokenInfo;
 
@@ -60,13 +60,13 @@ void NDR_SetTreeTokenInfoKeyword(NDR_TreeTokenInfo* tokenInformation, char* keyw
 void NDR_SetTreeTokenInfoToken(NDR_TreeTokenInfo* tokenInformation, char* token);
 void NDR_SetTreeTokenInfoLine(NDR_TreeTokenInfo* tokenInformation, size_t lineNumber);
 void NDR_SetTreeTokenInfoColumn(NDR_TreeTokenInfo* tokenInformation, size_t columnNumber);
-void NDR_SetTreeTokenInfoNodeNumber(NDR_TreeTokenInfo* tokenInformation, size_t nodeNumber);
+void NDR_SetTreeTokenInfoNodeNumber(NDR_TreeTokenInfo* tokenInformation, long nodeNumber);
 
 char* NDR_GetTreeTokenInfoKeyword(NDR_TreeTokenInfo* tokenInformation);
 char* NDR_GetTreeTokenInfoToken(NDR_TreeTokenInfo* tokenInformation);
 size_t NDR_GetTreeTokenInfoLine(NDR_TreeTokenInfo* tokenInformation);
 size_t NDR_GetTreeTokenInfoColumn(NDR_TreeTokenInfo* tokenInformation);
-size_t NDR_GetTreeTokenInfoNodeNumber(NDR_TreeTokenInfo* tokenInformation);
+long NDR_GetTreeTokenInfoNodeNumber(NDR_TreeTokenInfo* tokenInformation);
 
 NDR_TreeTokenInfo* NDR_GetTreeTokenInfo(NDR_TreeTokenInfoWrapper* tokenInfo, size_t index);
 NDR_TreeTokenInfo* NDR_GetLastTreeTokenInfo(NDR_TreeTokenInfoWrapper* tokenInfo);
